Create SCNGeometrySource from NSData to make use of floats clear
SCNVector3 in the simulator is a CGFloat which can be 32-bit or 64-bit,
whereas it’s a Float on a device. When running on a 64-bit simulator
this caused a failure as the vector data was incorrectly interpreted
as being made up of doubles.

This commit moves to creating its SCNGeometrySource instances from
an NSData which allows the fact the floats are being used to be
explicitly declared. This removes the “double not supported” error and
allows the app to run as expected.